hal: fix voip usecase selection for voice_tx path
Some VoIP clients may choose to open the input streams as soon as the phone
state is set to MODE_RINGTONE to allow lower start latency when the call
is connected. In such situation dev mode will not be still set to
MODE_IN_COMMUNICATION, hence removing condition to check for dev mode.

Change-Id: I38aeb137b2852bab6c3b90d4438179c375be946c